+++ packages/sdk-client/src/http/fetch/fetch-http-response.ts
@@ -1,0 +1,12 @@
+import { Response } from 'node-fetch';
+import { HttpResponse } from '../http-response';
+import { FetchHttpContentParser } from './fetch-http-content-parser';
+import { fromFetchHeaders } from './fetch-http-headers';
+
+export function toHttpResponse(nativeResponse: Response): HttpResponse {
+  return new HttpResponse(
+    nativeResponse.status,
+    fromFetchHeaders(nativeResponse.headers),
+    new FetchHttpContentParser(nativeResponse),
+  );
+}